finalisé l'API de saved posts

+++ app/Http/Controllers/Api/v1/ApiSavedPostController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers\Api\v1;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Post;
 use App\Models\SavedPost;
 
@@ -28,22 +30,17 @@ class ApiSavedPostController extends Controller
     public function store(Request $request)
     {
         $user = Auth::user();
+        $post = Post::findOrFail($request->post_id);
         
-        // si le post était déjà sauvegardé
-        $existing = SavedPost::where('user_id', $user->id)
-            ->where('post_id', $post->id)
-            ->first();
+        $savedPost = new SavedPost();
 
-        if ($existing) {
-            $existing->delete();
-        } else {
-            $savedPost = new SavedPost();
-            $savedPost->user_id = $user->id;
-            $savedPost->post_id = $post->id;
-            $savedPost->save();
-        }
+        // enregistre le post chez l'utilisateur connecté. donc pas besoin de la Policy, 
+        // car personne peut enregister un post chez qqn d'autre
+        $savedPost->user_id = $user->id;
+        $savedPost->post_id = $post->id;
+        $savedPost->save();
 
-        return $post;
+        return $savedPost;
     }
 
     /**
@@ -67,7 +64,10 @@ class ApiSavedPostController extends Controller
      */
     public function destroy(string $id)
     {
-        $user = Auth::user();
+        $savedPost = SavedPost::findOrFail($id);
+        
+        // seulement la personne qui a enregistré le post peut le désenregistrer
+        Gate::authorize('delete', $savedPost);
 
         $savedPost->delete();
